requirements.txt
opencv-contrib-python
vptree
imutils
streamlit
google-cloud-storage